Move channel fsm to decidated folder
Move all files related to the channel fsm to the fsm folder.
No functional changes in this commit.
