+++ eclair-core/src/main/scala/fr/acinq/eclair/NodeParams.scala
@@ -21,8 +21,9 @@ import fr.acinq.bitcoin.Crypto.PublicKey
 import fr.acinq.bitcoin.{Block, ByteVector32, Crypto, Satoshi}
 import fr.acinq.eclair.Setup.Seeds
 import fr.acinq.eclair.blockchain.fee._
-import fr.acinq.eclair.channel.Channel.{ChannelConf, UnhandledExceptionStrategy}
-import fr.acinq.eclair.channel.{Channel, ChannelFlags}
+import fr.acinq.eclair.channel.ChannelFlags
+import fr.acinq.eclair.channel.fsm.Channel
+import fr.acinq.eclair.channel.fsm.Channel.{ChannelConf, UnhandledExceptionStrategy}
 import fr.acinq.eclair.crypto.Noise.KeyPair
 import fr.acinq.eclair.crypto.keymanager.{ChannelKeyManager, NodeKeyManager}
 import fr.acinq.eclair.db._
